editor/libeditor/html/tests/test_bug290026.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=290026
-->
<head>
  <title>Test for Bug 290026</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=290026">Mozilla Bug 290026</a>
<p id="display"></p>
<div id="editor" contenteditable></div>

<pre id="test">
<script type="application/javascript">

/** Test for Bug 290026 **/
SimpleTest.waitForExplicitFinish();

var editor = document.getElementById("editor");
editor.innerHTML = '<p></p><ul><li>Item 1</li><li>Item 2</li></ul><p></p>';
editor.focus();

addLoadEvent(function() {

  var sel = window.getSelection();
  sel.removeAllRanges();
  var lis = document.getElementsByTagName("li");
  var range = document.createRange();
  range.setStart(lis[0], 0);
  range.setEnd(lis[1], lis[1].childNodes.length);
  sel.addRange(range);
  document.execCommand("indent", false, false);
  var oneindent = '<p></p><ul style="margin-left: 40px;"><li>Item 1</li><li>Item 2</li></ul><p></p>';
  is(editor.innerHTML, oneindent, "a once indented bulleted list");
  document.execCommand("indent", false, false);
  var twoindent = '<p></p><ul style="margin-left: 80px;"><li>Item 1</li><li>Item 2</li></ul><p></p>';
  is(editor.innerHTML, twoindent, "a twice indented bulleted list");
  document.execCommand("outdent", false, false);
  todo_is(editor.innerHTML, oneindent, "outdenting a twice indented bulleted list");

  // done
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
